+++ packages/elvis-colors/src/getThemeColor.ts
@@ -20,7 +20,6 @@ const getThemeColorObject = (label: ColorLabel, themeName: ThemeName): Color | n
     theme.state[label as keyof typeof theme.state] ??
     theme.text[label as keyof typeof theme.text] ??
     theme.background[label as keyof typeof theme.background] ??
-    theme.static[label as keyof typeof theme.static] ??
     null;
   if (!color) {
     console.error(`Color ${label} not found.`);