+++ packages/elvis-colors/src/themes/darkTheme.ts
@@ -211,14 +211,4 @@ export const darkTheme = {
       contrast: darkThemeColors['data-colors']['red-tomato'].contrastText,
     },
   },
-  static: {
-    'static-white': {
-      hex: darkThemeColors['primary-colors'].white.color,
-      contrast: darkThemeColors['primary-colors'].white.contrastText,
-    },
-    'static-black': {
-      hex: darkThemeColors['primary-colors'].black.color,
-      contrast: darkThemeColors['primary-colors'].black.contrastText,
-    },
-  },
 } as const satisfies Theme;